requirements.txt
pydantic-ai
feedparser
langdetect
keybert
textblob
sentence-transformers
python-dotenv 
nest_asyncio
python-telegram-bot
aiohttp
spacy>=3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz
hdbscan